Use common test node in da tests
