+++ nomos-services/data-availability/tests/src/lib.rs
@@ -1,9 +1,14 @@
 // Networking is not essential for verifier and indexer tests.
 // Libp2p network is chosen for consensus requirement, mixnet is ignored.
 
-#[cfg(all(test, feature = "libp2p", not(feature = "mixnet")))]
+#[cfg(test)]
+#[cfg(feature = "libp2p")]
 mod common;
-#[cfg(all(test, feature = "libp2p", not(feature = "mixnet")))]
+
+#[cfg(test)]
+#[cfg(feature = "libp2p")]
 mod indexer_integration;
-#[cfg(all(test, feature = "libp2p", not(feature = "mixnet")))]
+
+#[cfg(test)]
+#[cfg(feature = "libp2p")]
 mod verifier_integration;
